add binary search and some formatting

+++ quicksort.h
@@ -1,7 +1,8 @@
 #ifndef QUICKSORT_H
 #define QUICKSORT_H
 
-static void merge(int *left, int l_size, int *right, int r_size, int *arr, int arr_size);
+static void merge(int *left, size_t l_size, int *right, size_t r_size, int *arr,
+                  size_t arr_size);
 void merge_sort(int *arr, size_t l);
 
 #endif
